// client/src/api/models/InventoryItem.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from "../runtime";
/**
 *
 * @export
 * @interface InventoryItem
 */
export interface InventoryItem {
  /**
   *
   * @type {number}
   * @memberof InventoryItem
   */
  id: number;
  /**
   *
   * @type {number}
   * @memberof InventoryItem
   */
  survivorId: number;
  /**
   *
   * @type {string}
   * @memberof InventoryItem
   */
  itemType: string;
  /**
   *
   * @type {number}
   * @memberof InventoryItem
   */
  amount: number;
}

/**
 * Check if a given object implements the InventoryItem interface.
 */
export function instanceOfInventoryItem(value: object): value is InventoryItem {
  if (!("id" in value) || value["id"] === undefined) return false;
  if (!("survivorId" in value) || value["survivorId"] === undefined)
    return false;
  if (!("itemType" in value) || value["itemType"] === undefined) return false;
  if (!("amount" in value) || value["amount"] === undefined) return false;
  return true;
}

export function InventoryItemFromJSON(json: any): InventoryItem {
  return InventoryItemFromJSONTyped(json, false);
}

export function InventoryItemFromJSONTyped(
  json: any,
  ignoreDiscriminator: boolean,
): InventoryItem {
  if (json == null) {
    return json;
  }
  return {
    id: json["id"],
    survivorId: json["survivor_id"],
    itemType: json["item_type"],
    amount: json["amount"],
  };
}

export function InventoryItemToJSON(json: any): InventoryItem {
  return InventoryItemToJSONTyped(json, false);
}

export function InventoryItemToJSONTyped(
  value?: InventoryItem | null,
  ignoreDiscriminator: boolean = false,
): any {
  if (value == null) {
    return value;
  }

  return {
    id: value["id"],
    survivor_id: value["survivorId"],
    item_type: value["itemType"],
    amount: value["amount"],
  };
}